Grad-CAM backend/requirements.txt
numpy>=1.26.4,<2.0.0
opencv-python-headless==4.10.0.84
flask==3.0.3
gunicorn==22.0.0
pillow==10.4.0
tensorflow-cpu>=2.16.1